[QC-837] Apply policies only for data in a given period (#1403)
* [QC-837] Apply policies only for data in a given period

* Update config-test.yaml

* fixes for gh

+++ Framework/script/RepoCleaner/rules/none_kept.py
@@ -9,7 +9,7 @@ from Ccdb import Ccdb, ObjectVersion
 logger = logging  # default logger
 
 
-def process(ccdb: Ccdb, object_path: str, delay: int, extra_params: Dict[str, str]):
+def process(ccdb: Ccdb, object_path: str, delay: int,  from_timestamp: int, to_timestamp: int, extra_params: Dict[str, str]):
     '''
     Process this deletion rule on the object. We use the CCDB passed by argument.
     
@@ -18,6 +18,8 @@ def process(ccdb: Ccdb, object_path: str, delay: int, extra_params: Dict[str, st
     :param ccdb: the ccdb in which objects are cleaned up.
     :param object_path: path to the object, or pattern, to which a rule will apply.
     :param delay: the grace period in minutes during which a new object is never deleted.
+    :param from_timestamp: only objects created after this timestamp are considered.
+    :param to_timestamp: only objects created before this timestamp are considered.
     :param extra_params: a dictionary containing extra parameters (unused in this rule)
     :return a dictionary with the number of deleted, preserved and updated versions. Total = deleted+preserved.
     '''
@@ -29,22 +31,24 @@ def process(ccdb: Ccdb, object_path: str, delay: int, extra_params: Dict[str, st
     deletion_list: List[ObjectVersion] = []
 
     for v in versions:
-        if v.validFromAsDt < datetime.now() - timedelta(minutes=delay):
-            logger.debug(f"not in the grace period, we delete {v}")
-            deletion_list.append(v)
-            ccdb.deleteVersion(v)
-        else:
-            preservation_list.append(v)
+        if v.validFromAsDt < datetime.now() - timedelta(minutes=delay):  # grace period
+            logger.debug(f"not in the grace period")
+            if from_timestamp < v.validFrom < to_timestamp:  # in the allowed period
+                logger.debug(f"in the allowed period (from,to), we delete {v}")
+                deletion_list.append(v)
+                ccdb.deleteVersion(v)
+                continue
+        preservation_list.append(v)
 
-    # logger.debug("deleted : ")
-    # for v in deletion_list:
-    #     logger.debug(f"   {v}")
-    #
-    # logger.debug("preserved : ")
-    # for v in preservation_list:
-    #     logger.debug(f"   {v}")
+    logger.debug("deleted : ")
+    for v in deletion_list:
+        logger.debug(f"   {v} - {v.validFrom}")
 
-    return {"deleted" : len(deletion_list), "preserved": len(preservation_list), "updated" : 0}
+    logger.debug("preserved : ")
+    for v in preservation_list:
+        logger.debug(f"   {v} - {v.validFrom}")
+
+    return {"deleted": len(deletion_list), "preserved": len(preservation_list), "updated": 0}
     
 def main():
     ccdb = Ccdb('http://ccdb-test.cern.ch:8080')
